refactor: moved schema to its own file
